# Remove costly allocation from writePacketHeader by moving buffer into struct