My production code uses pcapgo.WritePacket to write header and packet data to a byte slice in the real time capture tight loop. While profiling this code, I discovered pcapgo.writePacketHeader's byte array seems to get allocated on the heap. I found by moving this into the Writer struct this allocation is not made saving approx. 30-40ns per call on my machine. While this won't matter much for post-capture processing, it does make a large impact during the capture itself. Moving this will make these write routines safe to call only from a single goroutine, but I can't see what use case could be had by calling from multiple anyway.

Before:
---------
    scott@saruman:~/go/src/github.com/google/gopacket/pcapgo$ go test -v ./... -run=^$ -bench=.
    goos: linux
    goarch: amd64
    pkg: github.com/google/gopacket/pcapgo
    BenchmarkWritePacket-8   	20000000	       116 ns/op
    PASS
    ok  	github.com/google/gopacket/pcapgo	2.510s

After:
-------
    scott@saruman:~/go/src/github.com/google/gopacket/pcapgo$ go test -v ./... -run=^$ -bench=.
    goos: linux
    goarch: amd64
    pkg: github.com/google/gopacket/pcapgo
    BenchmarkWritePacket-8   	20000000	        73.6 ns/op
    PASS
    ok  	github.com/google/gopacket/pcapgo	2.726s